+++ brainlets/girlsfrontline/static/girlsfrontline/js/doll-utils.js
@@ -159,6 +159,69 @@ const SPECIAL_DEFAULT_EQUIPS_NIGHT = { //numbers indicate ID of the equipment
 const SPECIAL_DEFAULT_EQUIPS_NIGHT_UNRELEASED = { //numbers indicate ID of the equipment
 };
 
+export const VALID_EQUIPS = [
+  [[4, 13], [6], [10, 12]],          //hg
+  [[10, 12], [6], [1, 2, 3, 4, 13]], //smg
+  [[5], [1, 2, 3, 13], [15]],        //rf
+  [[1, 2, 3, 4, 13], [8], [10, 12]], //ar
+  [[5], [1, 2, 3], [14]],            //mg
+  [[11], [7, 9], [1, 2, 3, 4]]       //sg
+]; 
+
+export const SPECIAL_VALID_EQUIPS = { //numbers indicate TYPE of the equipment
+  133: [-1, 5, -1], //6P62
+  208: [-1, 5, -1], //C-MS
+  178: [-1, 5, -1],  //Contender
+  72: [-1, -1, 18], //M1918
+  264: [-1, 39, 18], //M1918
+  35: [16, -1, -1], //Springfield
+  26: [19, -1, -1], //MP5
+  56: [[20, 21, 22], -1, -1], //AK-47
+  64: [[20, 21, 22], -1, -1], //Type56-1
+  96: [25, -1, -1], //UMP9
+  97: [25, -1, -1], //UMP40
+  98: [25, -1, -1], //UMP45
+  270: [25, -1, 49], //UMP45 mod3
+  113: [26, -1, -1], //9A-91
+  180: [-1, -1, 27], //Ameli
+  259: [-1, -1, 28], //M4A1 mod3
+  252: [-1, 30, -1], //M1911 mod3
+  268: [31, -1, -1], //IDW mod3
+  269: [-1, -1, 32], //Type64 mod3
+  258: [-1, 33, -1], //FN-49 mod3
+  44: [-1, 34, -1], //Kar98k
+  63: [35, -1, -1], //416
+  83: [-1, -1, 36], //MG3
+  41: [-1, -1, 37], //PTRD
+  256: [-1, 38, 23], //Mosin-Nagant mod3
+  38: [-1, -1, 23], //Mosin-Nagant
+  253: [40, -1, -1], //M1895 mod3
+  267: [41, -1, -1], //MP446 mod3
+  257: [-1, -1, 42], //SV-98 mod3
+  249: [43, -1, -1], //CLEAR
+  250: [44, -1, -1], //FAIL
+  66: [45, -1, -1], //FAMAS
+  251: [46, -1, -1], //SAA mod3
+  266: [-1, 47, -1], //Bren mod3
+  262: [48, -1, -1], //G3 mod3
+  60: [-1, -1, 51], //G41
+  254: [-1, -1, 52], //STEN mod3
+  255: [-1, 53, -1], //M14 mod3
+  263: [54, -1, -1], //G36 mod3
+  265: [-1, 55, -1], //LWMMG mod3
+  120: [-1, 56, -1], //MG4
+  7: [57, -1, -1], //Stechkin
+  289: [-1, 58, -1], //AS Val mod3
+  290: [-1, 59, -1], //StG44 mod3
+  291: [-1, -1, 60], //Micro Uzi mod3
+  296: [[61, 64, 65], [62, 63, 66], [68, 67]], //Jill
+  292: [69, -1, -1], //Dana
+  294: [-1, -1, 70], //Stella
+  295: [-1, -1, 71], //Sei
+  297: [-1, 72, -1], //Dorothy
+  293: [-1, -1, 73], //Alma
+};
+
 export default class dollUtils {
   /**
    * Initialize values.
