+++ brainlets/girlsfrontline/templates/girlsfrontline/sim.html
@@ -2049,6 +2049,7 @@
   </div>
   <script src="{% static 'girlsfrontline/js/damage-utils.js' %}" type="module"></script>
   <script src="{% static 'girlsfrontline/js/doll-utils.js' %}" type="module"></script>
+  <script src="{% static 'girlsfrontline/js/ui-utils.js' %}" type="module"></script>
   <script src="{% static 'girlsfrontline/js/sim.js' %}" type="module"></script>
 </main>
 {% endblock %}
